Added support to change ceph osd ownership
When upgrading from hammer to jewel the default owner should be
changed from root to ceph.  This requires a recursive chown
which can take sometimes weeks for big clusters.  The alternative
is to tell ceph to continue using root with a ceph.conf change.
I plan on supporting both options. The walkdir crate is supposed
to have nearly identical speed to find so I'm going to use that
instead of shelling out with Command.