Add config file settings for jewel
Adds a ceph.conf flag to use the existing /var/lib/ceph user
for daemons.  This kicks the can down the road for chowning the
ceph-osd files but that's preferable because it can take forever
to do. A heuristic could be added in the future for small osds to
just take the hit and change users on all the files.